feat(pwa): rewrite chat virtual scroll for smooth scrolling
Replace the Fenwick/spacer virtualization with a Telegram-style full-render
of the active segment plus anchor-delta scroll preservation, eliminating
estimated-height jitter on prepend and image load.
