+++ wetty-chat-mobile/src/components/chat/virtualScroll.legacy/fenwick.ts
