+++ wetty-chat-mobile/src/components/chat/virtualScroll/ChatVirtualScroll.module.scss
@@ -1,96 +1,91 @@
 .container {
   position: relative;
-  width: 100%;
   height: 100%;
+  width: 100%;
+  overflow-x: hidden;
   overflow-y: auto;
-  -webkit-overflow-scrolling: touch;
+  // Manual scroll preservation (anchor delta). Browser scroll anchoring is
+  // idempotent with it but inconsistent with flex-end, so disable.
   overflow-anchor: none;
-  background: inherit;
+  -webkit-overflow-scrolling: touch;
 }
 
 .flowContent {
-  position: relative;
+  display: flex;
+  flex-direction: column;
+  // Pin messages to the bottom when content is shorter than the viewport
+  // (matches Telegram `.messages-container`). When content overflows, it
+  // scrolls naturally from the top.
+  justify-content: flex-end;
   width: 100%;
   min-height: 100%;
-  overflow-anchor: none;
-  background: inherit;
 }
 
-.flowItem {
+.sentinel {
+  // Zero-height sentinel observed by IntersectionObserver to trigger loading.
   width: 100%;
+  height: 0;
+  flex-shrink: 0;
 }
 
-.rowContent {
+.row {
   width: 100%;
-  background: transparent;
-  transition: background-color 100ms ease-out;
+  flex-shrink: 0;
 }
 
-.rowContentHighlighted {
-  background: rgba(var(--ion-color-warning-rgb, 255, 196, 9), 0.18);
-  transition: background-color 100ms ease-out;
-}
-
-.dateRowContentHidden {
-  visibility: hidden;
-}
-
-.spacer {
-  width: 100%;
-}
-
-.boundaryRow {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 0.85em;
-  color: var(--ion-color-medium, #888);
-}
-
-.edgeHintRow {
-  position: absolute;
-  left: 0;
+.dateRow {
+  position: sticky;
+  top: 0;
   z-index: 2;
+  transition: opacity 0.3s ease;
+}
+
+// While scrolling, the stuck date is visible. After the user stops (~idle),
+// fade it out so it doesn't obscure messages (mirrors Telegram's
+// is-scrolling-messages → .stuck opacity:0).
+.container:not([data-scrolling='true']) .dateRow[data-stuck] {
+  opacity: 0;
+}
+
+.edgeHint {
   display: flex;
+  flex-shrink: 0;
   align-items: center;
   justify-content: center;
+  gap: 8px;
   width: 100%;
   font-size: 0.85em;
   color: var(--ion-color-medium, #888);
-  background: inherit;
   pointer-events: none;
 }
 
-.edgeHintTop {
-  top: 0;
+.edgeHintSpinner {
+  width: 18px;
+  height: 18px;
+  color: var(--ion-color-medium, #888);
 }
 
-.edgeHintBottom {
-  bottom: 0;
+.bottomPadding {
+  width: 100%;
+  flex-shrink: 0;
 }
 
-.stagingArea {
+.overlay {
   position: absolute;
   top: 0;
   left: 0;
-  width: 100%;
+  right: 0;
+  z-index: 2;
   pointer-events: none;
-  visibility: hidden;
-  z-index: -1;
 }
 
-.stagingItem {
-  width: 100%;
-}
-
-.overlayScrim {
+.loadingScrim {
   position: absolute;
   inset: 0;
   display: flex;
   align-items: center;
   justify-content: center;
+  background: var(--ion-background-color, #fff);
+  opacity: 0.85;
   z-index: 3;
-  background: var(--background, var(--ion-background-color-step-50, #ece5dd));
-  color: var(--ion-color-medium, #666);
-  font-size: 0.9rem;
 }
