+++ wetty-chat-mobile/src/store/index.ts
@@ -27,6 +27,7 @@ import chatsReducer, {
   projectChatMessagePatched,
 } from './chatsSlice';
 import pinsReducer from './pinsSlice';
+import highlightReducer from './highlightSlice';
 import userReducer, { fetchCurrentUser } from './userSlice';
 import { toMessagePreview, type MessagePreview, type MessageResponse } from '@/api/messages';
 import { messageAdded, messageConfirmed, messagePatched, messagesBulkDeleted } from './messageEvents';
@@ -315,6 +316,7 @@ const rootReducer = combineReducers({
   chats: chatsReducer,
   threads: threadsReducer,
   pins: pinsReducer,
+  highlight: highlightReducer,
   user: userReducer,
 });
